requirements.txt
ray[serve]==2.3.1
aiortc==1.3.2
aiohttp==3.8.5
diffusers==0.18.2
transformers==4.33.1
torch==2.0.1
torchvision==0.15.2
accelerate==0.21.0
Pillow==10.0.0
opencv-python-headless==4.8.0.76
pyyaml==6.0
requests==2.31.0
av==10.0.0
numpy==1.24.4
huggingface-hub==0.16.4
